test(img): migrate draggable
